+++ src/Phinx/Db/Table/Index.php
@@ -69,7 +69,7 @@ class Index
      * Sets the index columns.
      *
      * @param array $columns
-     * @return Column
+     * @return Index
      */
     public function setColumns($columns)
     {
@@ -165,8 +165,8 @@ class Index
         // Valid Options
         $validOptions = array('type', 'unique', 'name', 'limit');
         foreach ($options as $option => $value) {
-            if (!in_array($option, $validOptions)) {
-                throw new \RuntimeException('\'' . $option . '\' is not a valid index option.');
+            if (!in_array($option, $validOptions, true)) {
+                throw new \RuntimeException(sprintf('"%s" is not a valid index option.', $option));
             }
 
             // handle $options['unique']
@@ -180,5 +180,6 @@ class Index
             $method = 'set' . ucfirst($option);
             $this->$method($value);
         }
+        return $this;
     }
 }