+++ src/Phinx/Migration/Manager/Environment.php
@@ -129,13 +129,11 @@ class Environment
     /**
      * Executes the specified seeder on this environment.
      *
-     * @param MigrationInterface $migration Migration
-     * @param string $direction Direction
+     * @param SeedInterface $seed
      * @return void
      */
     public function executeSeed(SeedInterface $seed)
     {
-        $startTime = time();
         $seed->setAdapter($this->getAdapter());
 
         // begin the transaction if the adapter supports it